fix(matchidinput): the window can't show immediately when no network connection
Originally 'get today match' is running in a eventloop before the window show. So if there's no network connection, the eventloop will run until QNetworkReply throws a timeout error and the show event is blocked. The way to solve it is remove the eventloop and let it run in another thread created by the QNetworkAccessManager itself.